+++ spec/xcodeproj_ext_spec.rb
@@ -11,34 +11,34 @@ describe "Xcodeproj C ext" do
 
   it "coerces the given path object to a string path" do
     # @plist is a Pathname
-    Xcodeproj.write_plist({}, @plist)
-    Xcodeproj.read_plist(@plist).should == {}
+    Xcodeproj::PlistHelper.write_plist({}, @plist)
+    Xcodeproj::PlistHelper.read_plist(@plist).should == {}
   end
 
   if RUBY_VERSION >= '1.9' && !(defined?(RUBY_ENGINE) && RUBY_ENGINE == 'rbx')
     it "raises when the given path can't be coerced into a string path" do
-      lambda { Xcodeproj.write_plist({}, Object.new) }.should.raise TypeError
+      lambda { Xcodeproj::PlistHelper.write_plist({}, Object.new) }.should.raise TypeError
     end
   end
 
   it "raises if the given path doesn't exist" do
-    lambda { Xcodeproj.read_plist('doesnotexist') }.should.raise ArgumentError
+    lambda { Xcodeproj::PlistHelper.read_plist('doesnotexist') }.should.raise ArgumentError
   end
 
   it "coherces the given hash to a Hash" do
     o = Object.new
     def o.to_hash; { 'from' => 'object' }; end
-    Xcodeproj.write_plist(o, @plist)
-    Xcodeproj.read_plist(@plist).should == { 'from' => 'object' }
+    Xcodeproj::PlistHelper.write_plist(o, @plist)
+    Xcodeproj::PlistHelper.read_plist(@plist).should == { 'from' => 'object' }
   end
 
   it "raises when given a hash that can't be coerced to a Hash" do
-    lambda { Xcodeproj.write_plist(Object.new, @plist) }.should.raise TypeError
+    lambda { Xcodeproj::PlistHelper.write_plist(Object.new, @plist) }.should.raise TypeError
   end
 
   it "coerces keys to strings" do
-    Xcodeproj.write_plist({ 1 => '1', :symbol => 'symbol' }, @plist)
-    Xcodeproj.read_plist(@plist).should == { '1' => '1', 'symbol' => 'symbol' }
+    Xcodeproj::PlistHelper.write_plist({ 1 => '1', :symbol => 'symbol' }, @plist)
+    Xcodeproj::PlistHelper.read_plist(@plist).should == { '1' => '1', 'symbol' => 'symbol' }
   end
 
   it "allows hashes, strings, booleans, and arrays of hashes and strings as values" do
@@ -49,19 +49,19 @@ describe "Xcodeproj C ext" do
       'false_bool' => false,
       'array'  => ['string in an array', { 'a hash' => 'in an array' }]
     }
-    Xcodeproj.write_plist(hash, @plist)
-    Xcodeproj.read_plist(@plist).should == hash
+    Xcodeproj::PlistHelper.write_plist(hash, @plist)
+    Xcodeproj::PlistHelper.read_plist(@plist).should == hash
   end
 
   xit "coerces values to strings if it is a disallowed type" do
-    Xcodeproj.write_plist({ '1' => 1, 'symbol' => :symbol }, @plist)
-    Xcodeproj.read_plist(@plist).should == { '1' => '1', 'symbol' => 'symbol' }
+    Xcodeproj::PlistHelper.write_plist({ '1' => 1, 'symbol' => :symbol }, @plist)
+    Xcodeproj::PlistHelper.read_plist(@plist).should == { '1' => '1', 'symbol' => 'symbol' }
   end
 
   it "handles unicode characters in paths and strings" do
     plist = @plist.to_s + 'øµ'
-    Xcodeproj.write_plist({ 'café' => 'før yoµ' }, plist)
-    Xcodeproj.read_plist(plist).should == { 'café' => 'før yoµ' }
+    Xcodeproj::PlistHelper.write_plist({ 'café' => 'før yoµ' }, plist)
+    Xcodeproj::PlistHelper.read_plist(plist).should == { 'café' => 'før yoµ' }
   end
 
   xit "raises if a plist contains any other object type as value than string, dictionary, and array" do
@@ -77,7 +77,7 @@ describe "Xcodeproj C ext" do
 </plist>
 EOS
     end
-    lambda { Xcodeproj.read_plist(@plist) }.should.raise TypeError
+    lambda { Xcodeproj::PlistHelper.read_plist(@plist) }.should.raise TypeError
   end
 
   xit "raises if a plist array value contains any other object type than string, or dictionary" do
@@ -95,12 +95,12 @@ EOS
 </plist>
 EOS
     end
-    lambda { Xcodeproj.read_plist(@plist) }.should.raise TypeError
+    lambda { Xcodeproj::PlistHelper.read_plist(@plist) }.should.raise TypeError
   end
 
   xit "it raises if for whatever reason the value could not be converted to a CFTypeRef" do
     lambda do
-      Xcodeproj.write_plist({ "invalid" => "\xCA" }, @plist)
+      Xcodeproj::PlistHelper.write_plist({ "invalid" => "\xCA" }, @plist)
     end.should.raise TypeError
   end
 end